+++ jason/src/api/room.rs
@@ -1210,16 +1210,16 @@ impl EventHandler for InnerRoom {
     }
 
     #[inline]
-    async fn on_joined_room(&self, _: MemberId) -> Self::Output {
-        unreachable!("Room can't receive Event::JoinedRoom")
+    async fn on_room_joined(&self, _: MemberId) -> Self::Output {
+        unreachable!("Room can't receive Event::RoomJoined")
     }
 
     #[inline]
-    async fn on_left_room(
+    async fn on_room_left(
         &self,
         _: medea_client_api_proto::CloseReason,
     ) -> Self::Output {
-        unreachable!("Room can't receive Event::LeftRoom")
+        unreachable!("Room can't receive Event::RoomLeft")
     }
 }