+++ jason/CHANGELOG.md
@@ -45,7 +45,7 @@ All user visible changes to this project will be documented in this file. This p
         - Room management:
             - `Jason.init_room()` ([#46]);
             - `Room.join()` ([#46]);
-            - `Jason.close_room` ([#147]).
+            - `Jason.close_room()` ([#147]).
         - Ability to configure local media stream used by `Room` via `Room.set_local_media_settings()` ([#54], [#97], [#145]);
         - `Room.on_failed_local_media` callback ([#54], [#143]);
         - `Room.on_close` callback for WebSocket close initiated by server ([#55]);